backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0
